Added texture loading

+++ Game/src/Game.cpp
@@ -9,8 +9,8 @@ static void GlfwErrorCallback(int error, const char* description);
 static void APIENTRY GLDebugCallback(GLenum source, GLenum type, GLuint id, GLenum severity,
 	GLsizei length, const GLchar* message, const void* userParam);
 
-static constexpr unsigned int WindowWidth = 1280;
-static constexpr unsigned int WindowHeight = 720;
+static constexpr uint16_t WindowWidth = 1280;
+static constexpr uint16_t WindowHeight = 720;
 
 Game::Game() :
 	mRegistry(),
